// front/G-stock/src/app/services/models/produit.ts
import { Category } from './category'; // Adjust the import path if necessary

export interface Produit {
  idProduit?: number; // Optional, auto-generated by the backend
  nomProduit: string;
  description: string;
  prixUnitaire: number;
  codeAB?: string; // Optional, might be generated by the backend
  logo?: string; // Optional
  categorie?: Category; // Optional, depending on whether you need category details
}